requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
pendulum==2.1.0
pymongo==3.10.1
python-dateutil==2.8.1
pytzdata==2019.3
requests==2.23.0
six==1.14.0
StringDist==1.0.9
urllib3==1.25.8
Werkzeug==1.0.0
